chore: remove redundant code blocks from changes section
Eliminated multiple empty <code-changes> blocks to clean up
the codebase. This improves readability and maintainability
by reducing clutter and potential confusion for future
developers. No functional changes were made.
