+++ composeApp/src/androidMain/kotlin/com/example/seviya/feature/worker/WorkerDailyAppointmentsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.UI
+package com.example.seviya.feature.worker
 
 import android.content.Intent
 import android.net.Uri
