+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/landing/RoleCatalogScreen.kt
@@ -1,56 +1,30 @@
-package com.example.seviya.UI
+package com.example.seviya.feature.landing
 
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.animation.core.animateFloat
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.infiniteRepeatable
 import androidx.compose.animation.core.rememberInfiniteTransition
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.LocalIndication
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.collectIsPressedAsState
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.widthIn
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.composed
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.composed
 import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
@@ -67,6 +41,9 @@ import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.compose.animation.core.animateDpAsState
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.layout.offset
 import com.example.seviya.core.designsystem.theme.AvatarBlueSoft
 import com.example.seviya.core.designsystem.theme.BorderUltraSoft
 import com.example.seviya.core.designsystem.theme.BrandBlue
@@ -78,31 +55,28 @@ import com.example.seviya.core.designsystem.theme.SubtitleOnBlue
 import com.example.seviya.core.designsystem.theme.TextPrimary
 import com.example.seviya.core.designsystem.theme.TextSecondary
 import com.example.seviya.core.designsystem.theme.White
+
+// ✅ Tabler Icons (Compose Icons)
 import compose.icons.TablerIcons
-import compose.icons.tablericons.Briefcase
-import compose.icons.tablericons.Coin
-import compose.icons.tablericons.MapPin
-import compose.icons.tablericons.Search
-import compose.icons.tablericons.ShieldCheck
-import compose.icons.tablericons.User
+import compose.icons.tablericons.*
 import kotlinx.coroutines.delay
 
-private enum class RolesAdmissionTab { HOME, LOGIN, REGISTER }
-private enum class SelectedRoleAdmission { CLIENT, WORKER }
+private enum class RolesTab { HOME, LOGIN, REGISTER }
+private enum class SelectedRole { CLIENT, WORKER }
 
 /* ----------------------------- HELPERS ANIM ----------------------------- */
 
-private data class RoleAdmissionEnterAnim(
+private data class RoleEnterAnim(
     val alpha: Float,
     val offsetY: Dp,
     val scale: Float
 )
 
 @Composable
-private fun rememberRoleAdmissionEnterAnim(
+private fun rememberRoleEnterAnim(
     delayMs: Int,
     distance: Dp = 18.dp
-): RoleAdmissionEnterAnim {
+): RoleEnterAnim {
     var show by remember { mutableStateOf(false) }
 
     LaunchedEffect(Unit) {
@@ -128,34 +102,27 @@ private fun rememberRoleAdmissionEnterAnim(
         label = "enter_scale"
     )
 
-    return RoleAdmissionEnterAnim(alpha = alpha, offsetY = offsetY, scale = scale)
+    return RoleEnterAnim(alpha = alpha, offsetY = offsetY, scale = scale)
 }
 
 @Composable
-private fun RoleAdmissionAnimatedSection(
+private fun RoleAnimatedSection(
     delayMs: Int,
     modifier: Modifier = Modifier,
     content: @Composable () -> Unit
 ) {
-    val a = rememberRoleAdmissionEnterAnim(delayMs)
-
+    val a = rememberRoleEnterAnim(delayMs)
     Box(
         modifier = modifier
             .offset(y = a.offsetY)
             .alpha(a.alpha)
-            .graphicsLayer {
-                scaleX = a.scale
-                scaleY = a.scale
-            }
+            .graphicsLayer { scaleX = a.scale; scaleY = a.scale }
     ) {
         content()
     }
 }
 
-private fun Modifier.shimmerOverlay(
-    alpha: Float = 0.16f,
-    durationMs: Int = 1600
-): Modifier = composed {
+private fun Modifier.shimmerOverlay(alpha: Float = 0.16f, durationMs: Int = 1600): Modifier = composed {
     val inf = rememberInfiniteTransition(label = "shimmer")
     val x by inf.animateFloat(
         initialValue = -1f,
@@ -175,11 +142,7 @@ private fun Modifier.shimmerOverlay(
         val startX = x * w
 
         val brush = Brush.linearGradient(
-            colors = listOf(
-                Color.Transparent,
-                White.copy(alpha = alpha),
-                Color.Transparent
-            ),
+            colors = listOf(Color.Transparent, White.copy(alpha = alpha), Color.Transparent),
             start = Offset(startX, 0f),
             end = Offset(startX + shimmerW, h)
         )
@@ -198,10 +161,7 @@ private fun Modifier.bouncyClick(onClick: () -> Unit): Modifier = composed {
     )
 
     this
-        .graphicsLayer {
-            scaleX = s
-            scaleY = s
-        }
+        .graphicsLayer { scaleX = s; scaleY = s }
         .clickable(
             interactionSource = interaction,
             indication = LocalIndication.current,
@@ -212,14 +172,11 @@ private fun Modifier.bouncyClick(onClick: () -> Unit): Modifier = composed {
 /* ----------------------------- SCREEN ----------------------------- */
 
 @Composable
-fun RoleAdmissionCatalogScreen(
-    onGoHome: () -> Unit,
-    onGoLogin: () -> Unit,
-    onGoRegister: () -> Unit,
+fun RoleCatalogScreen(
     onPickClient: () -> Unit,
     onPickWorker: () -> Unit
 ) {
-    var selected by rememberSaveable { mutableStateOf<SelectedRoleAdmission?>(null) }
+    var selected by rememberSaveable { mutableStateOf<SelectedRole?>(null) }
 
     Scaffold(
         containerColor = White,
@@ -234,7 +191,7 @@ fun RoleAdmissionCatalogScreen(
             val bottomH = maxHeight - topH
 
             Column(Modifier.fillMaxSize()) {
-                RoleAdmissionHeaderAnimated(height = topH)
+                RoleHeaderAnimated(height = topH)
 
                 Column(
                     modifier = Modifier
@@ -245,12 +202,10 @@ fun RoleAdmissionCatalogScreen(
                     horizontalAlignment = Alignment.CenterHorizontally,
                     verticalArrangement = Arrangement.Center
                 ) {
-                    RoleAdmissionAnimatedSection(240) {
+                    RoleAnimatedSection(240) {
                         Text(
-                            text = "¿Cómo deseas ingresar?",
-                            style = MaterialTheme.typography.headlineLarge.copy(
-                                fontWeight = FontWeight.ExtraBold
-                            ),
+                            text = "Elige tu rol",
+                            style = MaterialTheme.typography.headlineLarge.copy(fontWeight = FontWeight.ExtraBold),
                             color = TextPrimary,
                             textAlign = TextAlign.Center
                         )
@@ -258,9 +213,9 @@ fun RoleAdmissionCatalogScreen(
 
                     Spacer(Modifier.height(8.dp))
 
-                    RoleAdmissionAnimatedSection(300) {
+                    RoleAnimatedSection(300) {
                         Text(
-                            text = "Selecciona el tipo de cuenta con el que\nquieres continuar",
+                            text = "Selecciona cómo usarás la plataforma para\ncontinuar",
                             style = MaterialTheme.typography.bodyLarge.copy(
                                 color = TextSecondary,
                                 lineHeight = 22.sp
@@ -271,32 +226,32 @@ fun RoleAdmissionCatalogScreen(
 
                     Spacer(Modifier.height(18.dp))
 
-                    RoleAdmissionAnimatedSection(360) {
+                    RoleAnimatedSection(360) {
                         Row(
                             modifier = Modifier
                                 .fillMaxWidth()
                                 .widthIn(max = 420.dp),
                             horizontalArrangement = Arrangement.spacedBy(16.dp)
                         ) {
-                            RoleAdmissionPickCardAnimated(
+                            RolePickCardAnimated(
                                 title = "Cliente",
-                                subtitle = "Ingresar al dashboard cliente",
+                                subtitle = "Busco servicios",
                                 icon = TablerIcons.User,
-                                selected = selected == SelectedRoleAdmission.CLIENT,
+                                selected = selected == SelectedRole.CLIENT,
                                 modifier = Modifier.weight(1f)
                             ) {
-                                selected = SelectedRoleAdmission.CLIENT
+                                selected = SelectedRole.CLIENT
                                 onPickClient()
                             }
 
-                            RoleAdmissionPickCardAnimated(
+                            RolePickCardAnimated(
                                 title = "Trabajador",
-                                subtitle = "Ingresar al dashboard trabajador",
+                                subtitle = "Ofrezco servicios",
                                 icon = TablerIcons.Briefcase,
-                                selected = selected == SelectedRoleAdmission.WORKER,
+                                selected = selected == SelectedRole.WORKER,
                                 modifier = Modifier.weight(1f)
                             ) {
-                                selected = SelectedRoleAdmission.WORKER
+                                selected = SelectedRole.WORKER
                                 onPickWorker()
                             }
                         }
@@ -307,10 +262,10 @@ fun RoleAdmissionCatalogScreen(
     }
 }
 
-/* ----------------------------- HEADER ----------------------------- */
+/* ----------------------------- HEADER CON MUCHAS ANIMACIONES ----------------------------- */
 
 @Composable
-private fun RoleAdmissionHeaderAnimated(height: Dp) {
+private fun RoleHeaderAnimated(height: Dp) {
     val inf = rememberInfiniteTransition(label = "header_inf")
 
     val gradT by inf.animateFloat(
@@ -368,7 +323,7 @@ private fun RoleAdmissionHeaderAnimated(height: Dp) {
         Box(
             modifier = Modifier
                 .size(260.dp)
-                .offset(x = (-110).dp, y = (-110).dp)
+                .offset((-110).dp, (-110).dp)
                 .background(BrandRed.copy(alpha = glowA), CircleShape)
         )
 
@@ -386,11 +341,11 @@ private fun RoleAdmissionHeaderAnimated(height: Dp) {
                 .padding(horizontal = 22.dp),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            RoleAdmissionLogoPillCompact()
+            RoleAnimatedSection(0) { LogoPillCompact() }
 
             Spacer(Modifier.height(14.dp))
 
-            RoleAdmissionAnimatedSection(80) {
+            RoleAnimatedSection(80) {
                 Text(
                     text = "Conectamos soluciones con\nnecesidades de forma segura",
                     color = White,
@@ -408,15 +363,15 @@ private fun RoleAdmissionHeaderAnimated(height: Dp) {
                 verticalArrangement = Arrangement.spacedBy(12.dp),
                 modifier = Modifier.widthIn(max = 320.dp)
             ) {
-                RoleAdmissionAnimatedSection(140) {
+                RoleAnimatedSection(140) {
                     Row(horizontalArrangement = Arrangement.spacedBy(12.dp)) {
-                        RoleAdmissionFeatureGlassTileSmall(
+                        FeatureGlassTileSmall(
                             icon = TablerIcons.Search,
                             label = "BUSCA",
                             modifier = Modifier.weight(1f),
                             float = floatY * 0.30f
                         )
-                        RoleAdmissionFeatureGlassTileSmall(
+                        FeatureGlassTileSmall(
                             icon = TablerIcons.Briefcase,
                             label = "OFRECE",
                             modifier = Modifier.weight(1f),
@@ -425,15 +380,15 @@ private fun RoleAdmissionHeaderAnimated(height: Dp) {
                     }
                 }
 
-                RoleAdmissionAnimatedSection(200) {
+                RoleAnimatedSection(200) {
                     Row(horizontalArrangement = Arrangement.spacedBy(12.dp)) {
-                        RoleAdmissionFeatureGlassTileSmall(
+                        FeatureGlassTileSmall(
                             icon = TablerIcons.ShieldCheck,
                             label = "CONFÍA",
                             modifier = Modifier.weight(1f),
                             float = floatY * 0.24f
                         )
-                        RoleAdmissionFeatureGlassTileSmall(
+                        FeatureGlassTileSmall(
                             icon = TablerIcons.Coin,
                             label = "GANA",
                             modifier = Modifier.weight(1f),
@@ -450,10 +405,7 @@ private fun RoleAdmissionHeaderAnimated(height: Dp) {
                 .fillMaxWidth()
                 .height(92.dp)
         ) {
-            RoleAdmissionPrettyWaveAnimated(
-                modifier = Modifier.fillMaxSize(),
-                baseColor = White
-            )
+            PrettyWaveAnimated(modifier = Modifier.fillMaxSize(), baseColor = White)
         }
     }
 }
@@ -461,7 +413,7 @@ private fun RoleAdmissionHeaderAnimated(height: Dp) {
 /* ---------------- PIEZAS UI ---------------- */
 
 @Composable
-private fun RoleAdmissionLogoPillCompact() {
+private fun LogoPillCompact() {
     Card(
         shape = RoundedCornerShape(999.dp),
         colors = CardDefaults.cardColors(containerColor = White),
@@ -493,30 +445,17 @@ private fun RoleAdmissionLogoPillCompact() {
 
             Text(
                 text = buildAnnotatedString {
-                    withStyle(
-                        SpanStyle(
-                            color = BrandBlue,
-                            fontWeight = FontWeight.ExtraBold
-                        )
-                    ) { append("Servi") }
-
-                    withStyle(
-                        SpanStyle(
-                            color = BrandRed,
-                            fontWeight = FontWeight.ExtraBold
-                        )
-                    ) { append("Ya") }
+                    withStyle(SpanStyle(color = BrandBlue, fontWeight = FontWeight.ExtraBold)) { append("Servi") }
+                    withStyle(SpanStyle(color = BrandRed, fontWeight = FontWeight.ExtraBold)) { append("Ya") }
                 },
-                style = MaterialTheme.typography.titleLarge.copy(
-                    fontWeight = FontWeight.ExtraBold
-                )
+                style = MaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.ExtraBold)
             )
         }
     }
 }
 
 @Composable
-private fun RoleAdmissionFeatureGlassTileSmall(
+private fun FeatureGlassTileSmall(
     icon: ImageVector,
     label: String,
     modifier: Modifier = Modifier,
@@ -526,10 +465,7 @@ private fun RoleAdmissionFeatureGlassTileSmall(
     val pulse by inf.animateFloat(
         initialValue = 1f,
         targetValue = 1.05f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(1800, easing = FastOutSlowInEasing),
-            repeatMode = RepeatMode.Reverse
-        ),
+        animationSpec = infiniteRepeatable(tween(1800, easing = FastOutSlowInEasing), RepeatMode.Reverse),
         label = "pulse_$label"
     )
 
@@ -537,34 +473,22 @@ private fun RoleAdmissionFeatureGlassTileSmall(
         modifier = modifier
             .height(78.dp)
             .offset(y = (-float).dp)
-            .graphicsLayer {
-                scaleX = pulse
-                scaleY = pulse
-            }
+            .graphicsLayer { scaleX = pulse; scaleY = pulse }
             .clip(RoundedCornerShape(20.dp))
             .shimmerOverlay(alpha = 0.12f, durationMs = 2000),
         shape = RoundedCornerShape(20.dp),
-        colors = CardDefaults.cardColors(
-            containerColor = White.copy(alpha = 0.12f)
-        ),
+        colors = CardDefaults.cardColors(containerColor = White.copy(alpha = 0.12f)),
         border = BorderStroke(1.dp, White.copy(alpha = 0.18f))
     ) {
         Column(
-            modifier = Modifier.fillMaxSize(),
+            Modifier.fillMaxSize(),
             verticalArrangement = Arrangement.Center,
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Icon(
-                imageVector = icon,
-                contentDescription = null,
-                tint = White,
-                modifier = Modifier.size(20.dp)
-            )
-
+            Icon(icon, null, tint = White, modifier = Modifier.size(20.dp))
             Spacer(Modifier.height(6.dp))
-
             Text(
-                text = label,
+                label,
                 color = White.copy(alpha = 0.78f),
                 style = MaterialTheme.typography.labelMedium.copy(
                     fontWeight = FontWeight.Black,
@@ -576,7 +500,7 @@ private fun RoleAdmissionFeatureGlassTileSmall(
 }
 
 @Composable
-private fun RoleAdmissionPickCardAnimated(
+private fun RolePickCardAnimated(
     title: String,
     subtitle: String,
     icon: ImageVector,
@@ -602,7 +526,7 @@ private fun RoleAdmissionPickCardAnimated(
         label = "icon_bg"
     )
 
-    val inf = rememberInfiniteTransition(label = "roleAdmission_float_$title")
+    val inf = rememberInfiniteTransition(label = "role_float_$title")
     val floatY by inf.animateFloat(
         initialValue = 0f,
         targetValue = if (selected) 8f else 0f,
@@ -621,9 +545,7 @@ private fun RoleAdmissionPickCardAnimated(
             .bouncyClick(onClick),
         shape = RoundedCornerShape(32.dp),
         colors = CardDefaults.cardColors(containerColor = bgColor),
-        elevation = CardDefaults.cardElevation(
-            defaultElevation = if (selected) 10.dp else 6.dp
-        ),
+        elevation = CardDefaults.cardElevation(defaultElevation = if (selected) 10.dp else 6.dp),
         border = BorderStroke(2.dp, borderColor)
     ) {
         Column(
@@ -656,9 +578,7 @@ private fun RoleAdmissionPickCardAnimated(
 
             Text(
                 text = title,
-                style = MaterialTheme.typography.titleLarge.copy(
-                    fontWeight = FontWeight.ExtraBold
-                ),
+                style = MaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.ExtraBold),
                 color = TextPrimary
             )
 
@@ -666,19 +586,17 @@ private fun RoleAdmissionPickCardAnimated(
 
             Text(
                 text = subtitle,
-                style = MaterialTheme.typography.bodyMedium.copy(
-                    color = TextSecondary
-                ),
+                style = MaterialTheme.typography.bodyMedium.copy(color = TextSecondary),
                 textAlign = TextAlign.Center
             )
         }
     }
 }
 
-/* ---------------- WAVE ---------------- */
+/* ---------------- WAVE ANIMADA ---------------- */
 
 @Composable
-private fun RoleAdmissionPrettyWaveAnimated(
+private fun PrettyWaveAnimated(
     modifier: Modifier,
     baseColor: Color
 ) {
@@ -686,10 +604,7 @@ private fun RoleAdmissionPrettyWaveAnimated(
     val a by inf.animateFloat(
         initialValue = 0.22f,
         targetValue = 0.38f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(2200, easing = FastOutSlowInEasing),
-            repeatMode = RepeatMode.Reverse
-        ),
+        animationSpec = infiniteRepeatable(tween(2200, easing = FastOutSlowInEasing), RepeatMode.Reverse),
         label = "wave_hi"
     )
